move key commands to archiver.keys
